requirements.txt
tensorflow-gpu
pygame